[KGEN] Use archive in MojoLLDB JIT, not module (#21849)
Instead of passing LLVM IR to MCJIT (in the form of an `llvm::Module`),
pass in a standalone archive instead (`llvm::object::Archive`).
However, note that:
* MCJIT requires an LLVM IR module for its initialization, even if that
  module is empty. Actual symbol lookup in this patch is handled
  entirely by the archive.
* The GDB registration listener needs to be disabled due to its
  asserting when the same archive is added twice. In Jupyter notebooks,
  this listener is shared across cell executions, and so is notified of
  standard library archives multiple times.

Although in and of itself this does not introduce a user-facing change,
it does fix a subtle correctness issue in which dependent archives are
not loaded in the REPL JIT. This manifests with on-demand compilation of
Mojo packages into standalone archives, which are then loaded into the
JIT -- but this on-demand comilation is implemented in a future commit,
and is tracked by #19983.

MODULAR_ORIG_COMMIT_REV_ID: eb322b5c1e1ed20266e3b4726d1b2ba1953e80a3