[KGEN] Move `compiledFor` to `PackageArchiveAttr` (#22360)
The `compiledFor` attributes on the `PackageOp` and `PackageLinkOp`
indicate what target the archive on those ops is built for. It's also
used as a sentinel -- if `compiledFor` is set on the op, then runtime
asserts ensure that `archive` must also be set.

That indicates that the target attribute should be part of
`PackageArchiveAttr`. This will also make sense when we begin storing
multiple precompiled archives on each of these ops -- we could store an
array of archives, each with a target attribute that indicates which
target the archive was built for.

Addresses a portion of #17326.

MODULAR_ORIG_COMMIT_REV_ID: 270ae7434e4c7a268d8a87b43cd2f5b0fc742935