[Kernels] Initial sketch of the elementwise GPU function (#22683)
This is an initial sketch of the elementwise GPU function. The function
does a dispatch into an empty kernel and so is not correct and cannot be
tested. It's an advancement over the existing non-existing
implementation though.

Subsequent PRs will fill this in.

RUNTIME_ORIG_REV_ID: e25c50e44a351899c44e6d543fe191e6fceea15a