[mojo] Compile missing package targets on-demand (#20802)
Closes #19983.

When a `.mojopkg` built on machine A for target X, is used on machine B
with a target Y, the `mojo` compiler determines that the `.mojopkg` does
not include code for target Y, and "falls back" to compiling the IR in
the `.mojopkg` for the host machine B.

Currently, the semantics of this fallback compilation behavior are
questionable: the `.mojopkg` includes pre-elaboration IR, and the `mojo`
compiler "merges" that IR into the current module being compiled. This
means that the current compilation's command-line options, environment
variable values, etc., are used to build the `.mojopkg`. But this
environment may be completely different from the one used to originally
build the `.mojopkg`.

This patch addresses the issue by instead performing the same
elaboration and static archive generation that `mojo package` does,
essentially compiling a package on-demand for any missing targets.

(Note that  due to issues with the MojoLLDB REPL, it maintains the old
behavior, where the pre-elaborated package module is added to the module
currently being built.)

MODULAR_ORIG_COMMIT_REV_ID: 9aaa85566d1a8693bfb922dadf013b5f393496de