[KGEN] Fix StackReuse incorrectly eliding copies of aliases (#22922)
Propagating known values across loads is used to elide loads, but it
doesn't work on loads from an alias because the pass does not track the
alias state and propagate it. Just make sure not to do this accidentally
for now to fix a crash.

MODULAR_ORIG_COMMIT_REV_ID: 33356d16275697ea493363d49abb71092c756a87