[CMake] [Cache] [KGEN] Regenerate CASID files only on input or KGENVersion change (#22417)
Before, due to quirks of the implementation, we would regenerate CASID
files every build, even if no inputs nor the KGENVersion had changed.
This was fast, but it is not free. In support of #19635, make the CASID
files have an explicit `DEPENDS` only on the inputs and KGENVersion.

Additionally, the existing implementation was written solely in CMake
using multiple targets and a CMake script, along with several temporary
files. While this works and it would be possible to make this change
while retaining this same structure, it is complicated and hard to
understand. The same logic can be written in Python, which I think makes
this part of the build process much easier to understand and change if
needed.

In support of #19635. Speeds up no-op `build` from 0.5s to 0.37s on my
machine (direct `ninja` invocation speeds up from 0.36s to 0.24s).

MODULAR_ORIG_COMMIT_REV_ID: 405a4987be1a78e702da20453398928236a0122c