[vscode] Improve how the RPC server works (#22985)
- Bump the internal version of the extension. This change is needed to
be able to install the extension internally using `vscode-init`.
Otherwise, vscode will stick to the version on the marketplace because
that has a higher number than what our package.json had. I haven't found
a better way to handle that for now. It seems that some companies have
two extensions with different names, one internal and one public. We
might need to do that as some point.
- Improve the documentation of the fields of the RPC server.
- Remove the host field of the RPC server, because it's always
localhost.
- When spawning servers, reuse the same server if it comes from two
different settings environments. This prevents some errors related to an
address already being used because the settings of two different
workspaces were pointing to the same configuration, which could happen
if the rpc config is set at the user level and not at the workspace
level. It turns out that you can't ask VSCode what's the origin of a
specific setting.
- Allow the same RPC server to use multiple tokens, which is needed for
reusability.
- Allow the RPC server to receive a moduleHomePath config from
vscode-debug.py, which is necessary to ensure that the correct dev SDK
is used. More information in the documentation of the package.json
- Unlink workspace folders from the RPC server. It turns out that this
is wrong because, if the setting is defined at user level or at a folder
level (not workspace, just folder), we'll never have a workspace folder,
and finding an SDK needs other means anyway.

Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: c472c84de7cc535a9ccfa3ab72cbccf92e4d58fb