[CUDA] Fix race condition on shutdown of CUDARuntime (#23083)
Though we use chains (both AsyncValueRef<Chain> and
AsyncValueRef<CUDAChain>) to make sure all CPU and CUDA values are ready
before they are used, those don't protect the model 'execute' block from
exiting with ready values while CUDA resources are still in use. That's
because finished CUDAAsyncValues may nevertheless still need to return
their stream and events to the pools, and, if required, free a CUDA
buffer.

I tried all sorts of elaborate and, in retrospect, bone headed ways to
fix this, but in the end the simplest is to simply make CUDAAsyncValues
hold an RCRef to its CUDARuntime so that its destructor will be called
only when no more CUDAAsyncValues are in flight.

With this turn the CUDA unit tests back on.

MODULAR_ORIG_COMMIT_REV_ID: f4dc940f4d026657a9e318898c29f59fb9fe3119