[LLCL] Allow out-of-order runtime allocation/deallocation (#23105)
The current implementation of runtime maintains a dense map of runtime
indices (`uint8_t`). This can be problematic if the order of
allocation/deallocation is not nested (i.e. allocate `1,2,3,4` and
deallocate `2` will cause fragmentation and not free any space).
The current PR maintains a vector of free indices which are used to
track the allocation. It is initialized with all indices in `uint8_t`
range. New allocations request an index from this vector and return the
index when done. This allows us to always be able to use the full range
of indices available even without nesting.
Addresses #23103.

MODULAR_ORIG_COMMIT_REV_ID: 7874361c78b0a5191cc6d85fbe5e8c2390161cde