[Stdlib] Throw an exception if tensor indexing mismatches with rank (#23081)
Before tensor indexing with length that did not match the rank caused
invalid access and behavior. This checks the rank matches and if not
throws an error. As a result of this, user code should handle the error
if it is thrown in tensor element access.

Closes #23078

EXTENSIBILITY_ORIG_REV_ID: 86dbfae9842876d3be380b98a5228553b29e7093