[LLDB] Rework how REPLResultRefType is handled in the type system. (#23121)
REPLResultRefType is an artificial type that is used to represent the
resultant variables from REPL expressions. Each resultant variable is
actually a pointer to its inner value.
We've been handling REPLResultRefTypes in a weird way: treating them
like references (like C++ & references), which ended up poluting our
type system with a lot of `dereference` calls. This is actually wrong
because this is not a reference type, and I assume we'll have it at some
point.
I'm now handling it as a regular type with a custom visualizer (aka
synthetic formatter). I had to improve the wrapping type formatter to
account for this case, which is a good idea anyway.
Besides that, the only two places where we need especial handling for
REPLResultRefTypes is in when computing its byte size (we cast it as a
pointer) and where we get its display name.
This code is very well covered by our existing tests.

---------

Co-authored-by: Brian Gesiak <brian@modular.com>
MODULAR_ORIG_COMMIT_REV_ID: c6b1f415f3421147c1bba7eb83f683604fb3b771