[LLDB] Revisit unimplement methods in the type system (#23238)
The type system is huge and has a lot of methods that we haven't
implemented. I'm doing a pass through all of them, adding simple
implementations to the ones that are trivial, fixing minor things, and
adding comments on all of them. This will serve for the upcoming work of
dwarf -> mlir translation, which uses the type system.

MODULAR_ORIG_COMMIT_REV_ID: 7929564169f4bb014b2cadb5e4d1c896fe11ece3