[All] Split up mt-multi-tenancy from API tests (#23228)
Split up the multi-tenancy feature exercised through `mt` with the core API
end-to-end tests.

This is a better separation of concerns, as there will be more API end-to-end
tests and we don't want them to litter their complexity with the one mt
end-to-end test for the multi-tenancy feature.

As an added bonus, since there are two separate CMake targets now, they can run
in parallel, which will speed up pre-submit.  The reason for this is that the mt
multi-tenancy feature is slow (due to downloading full models and executing
them), so the core API end-to-end CMake target can run in parallel on a
separate CPU.  It also gives us the ability (which I plan on doing in a future
PR) of moving this mt-multi-tenancy test over to post-submit as it will
probably rarely/never break, so it's overkill to run on pre-submit.  Now that we
have the infra in place for that though, we can enable that in the future
easily.

Other enhancements we may want to consider are moving the `concurrent.api` test
into its own stress-test category: it's a bit overkill to run that on every
pre-submit change.

MAX_INTEGRATION_TESTS_REV_ID: 6929448bb3092ace8a43ca8b9139870586c632db