[Kernels] Split the GPU module into separate files (#23328)
This splits the GPU module into multiple files. This makes it easier to
add components and not have one mega file.

We were also not running the right test due to a type, so this fixes
that as well.

ALGORITHM_ORIG_REV_ID: 04134d03256b171bc561ddbf0658d5640a0148ff