[Mojo][NFC] Add (non load bearing) argument passing kind storage to `FnMetadataAttr` (#23352)
This patch modifies the attribute and build methods, but keeps it from
being printed with the `lit.func` op, so that this change is easier to
review in isolation. The attribute is always filled with positional-only
kinds, which is going to be the correct behavior for closure emission
anyway, but will only be tested once the actual IR printing/parsing is
modified for `lit.func`.

MODULAR_ORIG_COMMIT_REV_ID: e5711c4b1ab010a715fd8ecd20fea0c95ac2ec4d