[Mojo] Encode argument passing kinds in `lit.func` and `lit.signature` syntax (#23176)
The IR assembly format of `lit.signature` relied on some hacky counting
of positional-only arguments. This patch modifies this so that the
argument passing kind is encoded in the IR using a similar syntax to
that of Python/Mojo: `*` to mark the start of keyword-only arguments,
and `|` to mark the end of positional-only args (because the `AsmParser`
has no function to parse a `/` yet).

As a consequence, the parser and diagnostic printer of function types in
Mojo can no longer rely on missing names encoding positional-only
passing kind. This patch modifies this logic to allow specifying a
function type with all arguments being positional-only (common use case
in the standard library and kernels) by omitting _all_ argument names.
That is, the following are equivalent:
```mojo
alias a = fn (Int, Int, String) -> Int
alias a = fn (Int, Int, String, /) -> Int
```

The following fails because name omission is only allowed if _all_
argument names are missing:
```mojo
alias b = fn (Int, b: Int) -> Int
```

Otherwise, the function signatures require `*` and `|` analogously to
function declarations, and are allowed to omit positional-only arguments
to omit the argument name. For example:
```mojo
alias b = fn (Int, /, b: Int) -> Int  # this works, because the '/' is explicit
alias c = fn (Int, /, String) -> Int  # fails, the second argument is pos-or-kw, must have name
```

MODULAR_ORIG_COMMIT_REV_ID: 10de475852fff6c39009d14c2d4bc93e9f470ee8