[Kernels] Add a test case for repeated values in algorithm.sort._quicksort (#23615)
Previous implementation before #23560 had a bug if comparator isn't
strictly equal it will not make progress (swap same indicies). The fix in the #23560 is legit but the implementation is still not ideal as it does more `cmp` that it should -- I will file a separate issue for that.

ALGORITHM_ORIG_REV_ID: 10a795f841fe98fd83e12c324db16a2b12805b51