[Telemetry][TF] Basic op and fallback logging (#23752)
Similar to the existing stats reporting done during the EAP for TF, plumb
through a telemetry context through to the lowering of TF to MO and emit the
same information covering things like:
- total op count
- amount of ops that went to fallback
- fallback op info (op signature, in/out types, attributes)

It wasn't clear whether we also want the non-fallback op histogram emitted as
telemetry or not.  In any case, this is all very experimental and subject to
change.  This is not the final telemetry data model we'll end up for ops; treat
this as a proof-of-concept just to enable others working on telemetry this
milestone.

We'll want to iterate on this end-to-end with defining what the SQL would look
like over different ways we choose to emit the op metrics to make life easy on
the data collection side in the infra repo.

Example logs as seen from otel-collector:

```
SeverityText: INFO
SeverityNumber: Info(9)
Body: Str()
Attributes:
     -> total_op_count: Int(4)
     -> fallback_op_count: Int(2)
     -> event.domain: Str(modular)
     -> event.name: Str(tf.stats)

SeverityText: INFO
SeverityNumber: Info(9)
Body: Str()
Attributes:
     -> Identity(tensor<1x2xf32>) -> (tensor<1x2xf32>) {"force_fallback"=unit}: Int(2)
     -> event.domain: Str(modular)
     -> event.name: Str(tf.stats.fallback)
```

Note:
- Still no easy way to have e2e tests with telemetry - see
  [Internal link]

Related to [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 0891c06e0b145520cb8daf2fe6060c011e2534d4