[mojo-lang] Pass copy-captured values into impl ctor by borrow (#23864)
Instead of owned. Passing as owned causes a copy to be emitted when the
impl is instantiated and a move to be emitted inside the constructor. By
passing as a borrow, only a single copy is emitted inside the
constructor, eliding one temporary variable.

MODULAR_ORIG_COMMIT_REV_ID: 14d3f5c5304d03bed187485aa1a5d48adfc93fcf