[LLDB] Add a dwarf parser for basic mojo programs (#23825)
This is the initial implementation of the dwarf parser that will power
our debugger experience. It's able to create decls and types of mlir
floats and ints, as well as mojo modules, structs and functions.
The dwarf parser class is used by the main LLDB parser via a not very
clear interface to convert dwarf entries into decls and types. The
actual code that creates the mlir entities is in the type system. This
process reminds me a bit of the mojo parser, but this one is extremely
simple and needs to account for some abnormalities, like objects without
names, duplicated struct entries, and some other issues that are being
tracked and will be resolved soon.

The resultant test shows the parser in action, where it's able to
reconstruct some basic variables and print their contents.

MODULAR_ORIG_COMMIT_REV_ID: aeb18ebfe079d599af739a2427eeebdbcc275c9b