[mojo-lang] Flatten single-field debuginfo structs (#23974)
Fixes #23484 (temporarily until better fix for #23914).

Since we don't have DWARF expression support yet, we cannot encode a
different type in the debuginfo for a decl than what the type actually
is in the IR.

This PR temporarily solves this issue by flattening structs in debuginfo
when we do the same to the IR type. This way the two stay in sync (at
least for the optimizations we know are triggering issues). The only
inconvenience is the type shown in the debugger won't be exactly what
the user declared.

A more proper fix utilizing dwarf expressions will come soon for #23914.

---------

Co-authored-by: Jeff Niu <jeff@modular.com>
MODULAR_ORIG_COMMIT_REV_ID: d7d0b03d2ab13a7f474a954b742cde170e0e393e