[Config] Add MODULAR_REDACT_VERSION_REVISION CMake option (#24331)
In this configuration, we do not include any information about the Git
revision in binaries, instead saying `<redacted>`. This is not
recommended in most cases -- this is critical diagnostic information
when we do need to diagnose problems. But, it dramatically speeds up
iteration: Without this, we have to re-link all executables. Worse,
because "all executables" includes the Mojo compiler, this means all
Mojo code needs to be recompiled. We essentially do not get any type of
incremental compilation across commits when the revision is included.

While we should continue to include revision by default, in certain
development use cases where commits are common, it would be good to have
a way to turn this off to avoid wasted time rebuilding.

MODULAR_ORIG_COMMIT_REV_ID: 6f627b48017cc521b683836679750bce4f7e0e5a