[Stdlib][Kernels][GPU] Make Pointer/DTypePointer address_space aware (#24160)
This makes the Pointer and DTypePointer address_space aware, which means
that we can delete the DevicePointer and DeviceDTypePointer types from
the GPU module. While I am at it, this changes the underlying
implementation of DTypePointer to be Pointer (eventually we can just
write this as just an alias).

GPU_ORIG_REV_ID: 63499aea36002ede309a6ca18fa244b7d8375226