[KGEN] Finalize package bytecode reader on error (#24505)
Closes #24437.

The issue is that the `mlir::BytecodeReader` must have `finalize` called
before it's destroyed, as it states in its documentation: "The
`finalize` method must be called before destruction." So, this patch
adds calls to `finalize` along all error paths.

There's already one such call to `finalize`, in the case that the pass
is unable to find a function in the imported package bytecode. This
patch copies that approach to the other error paths.

MODULAR_ORIG_COMMIT_REV_ID: 7907b9f771965c5cde445a7742758a45647f2856