[KGEN] Un-revert "Packages include 0-N precompiled targets" (#24537)
This originally landed as #23278, while the Mojo SDK post-install
self-tests were disabled (see MOOPS-9), and was then reverted in #24155.
The problem was with the condition being used to determine whether to
insert a `kgen.link` op -- one would be necessary only when a
pre-compiled package archive had been used, or when a package archive
had been compiled on-demand. But the condition was only true when a
pre-compiled archive was available, and not in the compiled on-demand
case.

This re-landing now includes the correct conditional -- and, if there
are any remaining problems, they'll be caught immediately post-submit,
since the post-install self-tests are running again. I also tested with
[Internal link] and
everything appears successful -- that stack being tested also includes a
change to work around #24159 and trigger self-tests before they're
landed to main.

Closes #24436.
MODULAR_ORIG_COMMIT_REV_ID: d6be7e90b60988d5ba93c55b7ca96faaa2eae92b