[Support] Fix `driver-tblgen` compilation in C++20 mode (#24566)
When compiling in C++20 mode, we get a hard error in building the
`driver-tblgen` tools.  This is because, in C++20, the rules about
incomplete types and allocators changed a bit with regard to ODR-use
and constexpr.

Specifically, when the constructor is defined, Clang tries to figure out whether
it can construct the `CommandOptionGroup` at compile-time.  To do this, it needs
to evaluate constructors (and destructors) for compile-time allocated memory,
which in our case, involves `std::vector<CommandOption>::~vector()`.

In our case, `CommandOptionGroup` contains a reference to a `CommandOption`,
which is incomplete (forward declared at the top of this header file).  This
makes us run into issues when actually the implicitly-generated destructor for
this class is checked, since it would require deallocating memory for the
`options` non-static data member.  But this won't work, of course, because we
don't even know the layout or size of `CommandOption` since it is an incomplete
type.

The easiest way to fix this is to move the constructor out-of-line and defined
in the C++ file where `CommandOption` would be fully defined at this point.
We can't just move `CommandOptionGroup` to be fully defined after
`CommandOption` since `CommandOption` has a friend function to a member
function of `CommandOptionGroup`.

MODULAR_ORIG_COMMIT_REV_ID: 4a0955ed3f6fce4f2de81f8d0822be5cb41a7473