[Support] Fix driver `printHelp` in C++20 mode (#24567)
Our td files may contain unicode characters such as the fire emoji, which is
encoded as `char8_t` - notably, a distinct type from `char` as of C++20.  When
tablegen is run, the td files are encoded as a big `char8_t[N]` array now
instead of `char[N]` array.  In turn, we were expecting these arrays to be
implicitly convertible to a `Twine` we use to print to the console before
exiting, when a user requests `--help` from a driver (such as `mojo`).

Fix this by adding an overload of `printHelp` that takes a `u8string_view` and
just constructs a `std::string_view` out of it.  Change the `printHelp` to just
take a `std::string_view`.  There's no interface requirement that we need a
`Twine` since we never concatenate strings or anything here - just immediately
print it. This saves an explicit construction of a `Twine` from a
`string_view` while we're here.  Since C++17 is the default still, guard the
`u8string_view` overload for only when we're in C++20 mode since that is a
C++20-or-later only library type.

MODULAR_ORIG_COMMIT_REV_ID: 117b515c3daba076ffb4e27b92373974ad47c456