[Support] change getHostFeatures to always use getFeaturesFromClang (#24655)
Change the helper to get the list of host features to always go through
the clang driver path instead of using the llvm helper. This is needed
for the ARM64 feature detection of dotprod/i8mm features.

More context:

The ARM64 version of llvm::sys::getHostCPUFeatures on Linux relies on
reading /proc/cpuinfo to extract CPU feature strings. The set of strings
that LLVM checks for is fairly limited and misses the dotprod/i8mm
strings I need. The Windows on ARM path also has a different code path
with even more limited support. This is all because the registers to
read feature info are not user mode accessible.

We already hit that the Apple M1 path for llvm::sys::getHostCPUFeatures
returns nothing, so the clang driver path was added. This change makes
the stack always go down this path.

I verified on a c6i.4xlarge (icelake-server) system that the clang path
returned more feature strings except for "+64bit" which is not used in
the code and can be derived from other target info.

MODULAR_ORIG_COMMIT_REV_ID: 2b91f751bc6d29dc9c416c96aa7f118e35172c06