[KGEN] Give the `?` syntax to UnboundAttr, but keep it for MO (#24684)
`#kgen.unbound` is about to become a lot more common in the IR, and it
makes sense to use `?` as the syntax for it. However, it conflicts with
`?` already given to `UnknownAttr`. KGEN still uses it to represent
undefined values, but not very commonly, so downgrade the syntax to `*?`
instead.

For MO, UnknownAttr is used very commonly in the shape representation,
so add some extra logic to keep the syntax as-is.

MODULAR_ORIG_COMMIT_REV_ID: 700d9b8a0a268284f62b0cb832aced83f93ffd68