[mojo-lang] deep check register-passability for POP::VariantType (#24653)
This is required because functions with `raises` wrap their return type
in a `POP::VariantType` with an error type. But if that return type is
parametric, a user can pass in a memory-only type. When this happened,
the check for register passability for the wrapped type would be
register passable and the type would not be rejected from being passed
in as a parameter. Until we delay register passability checks to the
elaborator, we need a hack like this.

Fixes [Internal link]

---------

Co-authored-by: Laszlo Kindrat <laszlo@modular.ai>
MODULAR_ORIG_COMMIT_REV_ID: 2a333d9e18c096aaa55287ac1ed884d63d4244c5