[Stdlib] Use 32bit arithmetic when computing index for shared memory (#24833)
Shared memory is not big, so a common optimization is to use int32
to perform indexing when using shared memory. Note that the ndbuffer
dimensions are still in int (so there will be casts). Future PRs will
explore storing the sizes of NDBuffer as UInt32 if we are in shared
memory address space.

BUFFER_ORIG_REV_ID: 03300ca56b7ac257bab87d93e2787e65858ba53f