[KGEN] Restrict the alnum mangling to just the NVPTX target (#24840)
In retrospect, this should've been how it was done to begin with, as it
would've saved lots of pain on the debugger bringup. However, originally
I did not do this because I didn't want the linkage name to change based
on the target. Of the two evils, this is probably the lesser until we
sort out this.

MODULAR_ORIG_COMMIT_REV_ID: eee54c653f2109610af8398577a29fa699b38c4e