[Support] Add ContainerReadCallbackAdaptor for HTTP requests. (#24649)
A lot of HTTP requests want to send the contents of a C++ buffer in the body, and the logic is a little fiddly. This patch adds an adaptor that can be used to adapt a container of bytes to the type required for the read callback.

MODULAR_ORIG_COMMIT_REV_ID: 7cf60e4de2ebaa031dd14d31481c759d1da54373