[mojo-lang] Make the BindTypeAttr API consistent (#24899)
I was debating with myself whether the attribute's values set should be
the full set of the type or just the set of new bindings. The current
implementation is a mix of both, so standardize on the latter. This
means BindTypeAttr takes responsibility for interleaving the new
bindings with the existing ones.

MODULAR_ORIG_COMMIT_REV_ID: bc527a3839d23db46ba45b7d57390cada99c7c9d