[mojo-lang][checklifetimes] Fix lifetime checking in obscure case (#24928)
Lifetimes are not trackable for register-passable trivial types -
the parser intentionally does not emit logic that would allow this
to happen in order to keep the IR compact.  That said, the
CheckLifetimes pass was using "absence of a destructor" as a proxy
for whether the type was trivial.  These are similar but different
things.

Fix the code to check for whether the type is trivial, allowing us
to correctly diagnose the testcase.

This fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 8a33084c8d78aabf95e641366195c8697d9316ed