[mojo-lang] Warn about pointless transfer on trivial types. (#24934)
Transfer on trivial register BValues and RValues will never do anything
- you
can't end the lifetime of such a construct. You *can* end the lifetime
of an
LValue (this marks it uninitialized) and CheckLifetimes supports that,
but we
should generate a warning for obviously pointless code. This was
discovered
as part of the work to look into
[Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 3456e6a6fc79c53037fb9f93d9d4319d5726e8b5