[Support] Aggregate the results from microbenchmarking better (#24873)
Changes the summarization code so that it only summarizes statically
significant entries. Those are entries that either exceed the min
runtime, are greater than the batch size the user asked for, or are in
the upper 10% of the runs.

Also remove metrics which are footguns and are no longer valid with this
filtering methodology

MODULAR_ORIG_COMMIT_REV_ID: c0588848ea085d4d873916b65b2bcc7196a3f1e6