[mojo-lang] For __mlir_op multi-return _type, use tuple instead of list literal (#24935)
In post-merge discussion of [Internal link] it was decided that we should take multiple return types as a tuple instead of a list. This implements the change.

One unexpected thing in this PR is the removal of a test case from `metatypes.mojo`.  This change broke the test because it relied on `--mojo-disable-builtins`, and this code path requires access to the TupleLiteral builtin type.  We decided that enough other things should test the code path the test was checking that it could just be removed.

MODULAR_ORIG_COMMIT_REV_ID: 43c2347c6860e648b1cab6510b920930e6fec619