[mojo-lsp] Provide documentation for import auto-completions (#24817)
This involves partially parsing in the module or package from the filesystem
so that we can build the markdown documentation. Most of this PR is
restructuring the ParserDriver to allow for parsing packages/modules to
support this use case.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: b3fb80586ac66c08098ea27c132fe54679c1d989