[mojo-lldb] Fix asan issue with archive buffer in non-cached builds (#24955)
When passing a generated archive to a JITExecutionUnit we use a
memory buffer constructed from a BufferRef, but don't ensure the
BufferRef is kept alive. We've currently been getting lucky here because
the in-memory cache backend takes a reference to the generated buffer
and keeps it alive, but this blows up when we don't cache. This was
uncovered in #24451, which changes the way we handle the in-memory
caching.

To avoid these footgun issues, a new OwningBinary class was added that
contains a binary and a backing BufferRef (an analog of the llvm class,
but for our buffer class).

MODULAR_ORIG_COMMIT_REV_ID: b4ba49eb70d80129bc61f668ce8ce5cac6faff68