[Kernels][GPU] Do not unload the CUDA module because context delete w… (#24969)
…ill do that

This changes the code so that cached modules are not unloaded before
program exit. The reason is that the context is destroyed before that
happens and the context destruction also unloads all the modules, so the
unload becomes redundant.

With this I get a clean result from the compute santizer

```
$ mojo build -parsing-stdlib Kernels/test/gpu/elementwise.mojo -o elementwise; /usr/local/cuda/bin/compute-sanitizer ./elementwise
ninja: no work to do.
========= COMPUTE-SANITIZER
== run_elementwise
42.0
43.0
43.0
44.0
44.0
45.0
43.0
44.0
44.0
45.0
45.0
46.0
========= ERROR SUMMARY: 0 errors
```

Closes [Internal link]

GPU_ORIG_REV_ID: b3eb4da874eddd977bfe71a8467b49b3afd78ab9