[mojo-lang] Fix no-`__exit__` context mgr problems with early exit (#24961)
As a Python extension, Mojo allows one to define context managers
that lack a `__exit__` method.  Mojo still needs to extend the
lifetime of the relevant value across the body of the with statement
though, so it emits a `lit.ownership.use` at the bottom of the block.

The problem with this is that early returns due to return/break etc
didn't work right: they reported the lit.ownership.use as being
unreachable, and they also didn't extend the lifetime.  Fix this by
emitting the lit.ownership.use into the finally block of the generated
code instead of at the bottom of straight-line code.

This fixes [Internal link] but the
externally reported issue hits another bug which I'll fix as a followup.

MODULAR_ORIG_COMMIT_REV_ID: ffe4d9ddda96165a7627102fb2fa9b0cd48dd2aa