[Kernels][Mogg] Make sure to rebuild mogg mlirbc file if kernels change (#24991)
When changing a kernel that should invalidate the mlirbc mogg file that
is created and trigger a rebuild. That's not what's happening today and
it causes massive uncertainty in terms of what a person is debugging. A
common workaround is to touch the MOGG cmake file before rebuilding, but
that requires extra builds that are not necessary.

This is the n-th attempt at doing this, but this appears to work.

MODULAR_ORIG_COMMIT_REV_ID: 008147feb8ebfcd5229b34dc687ed09adcdbc978