[LLDB] Render nicely booleans (#24820)
This adds the @lldb_formatter_wrapping_type decorator to Bool and adds the necessary type summary providers so that bools can be rendered as True or False in the REPL.
This patch should have been really small, but we needed the handlers for summaries for REPL results types and for the types with the decorator.

Sadly this doesn't work in a regular debug scenario because dwarf is wrong for bools. See [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: f6bafef475dc5950152543b8122204418a1ffced