[LLDB][DWARF] Use the new SourceName for generating function display names (#25055)
This uses the new `DebugInfo::SourceNameAttr`, which is emitted for
functions, as the source of information to generate function display
names that appear in frames, e.g. in stack traces.
There are few things that need to be worked out, like parameters in
nested functions, and adding function arguments in the non-leaf
functions, but this shows that we are in the right track.

MODULAR_ORIG_COMMIT_REV_ID: e6287a230874928d33e7e0521dfe753c884bf386