[KGEN] Correctly handle index reference depth remapping in many places (#25103)
This was a real pain in the neck to track down, especially since the
error from the parser is not informative, and anything with parameters
and deeply nested attributes is hard to track down.

This really boiled down to `upbindApplyResult` and `upbindValue` not
correctly adjusting the depth of index references that escaped and
referenced an outer signature when substituting values.
`mlir::AttrTypeReplace` doesn't know how the correctly handle this,
since the depth has to be tracked. Use `IndexRefRemapper`, which does
have the correct logic, to do depth adjustment instead.

Fixes [Internal link]
Fixes [Internal link]

And probably fixes other issues too.

MODULAR_ORIG_COMMIT_REV_ID: 9ea3ca09905d0c6098502695198858e48cdc6068