[KGEN] Properly generate debuginfo for GPU stubs (#25136)
The location used for the generate code was the one of the sliced
function, which could contain debuginfo with parameter references. For
the capture arg stub, just nuke out debuginfo, and for the instantiate
stub, create a subprogram, because the callee is expected to be inlined.

Fixes [Internal link]

GPU_TEST_ORIG_REV_ID: 2bb4e616bd6bf28b974e7f47071c19b21008d2ff