[mojo-lsp] Change diagnostic waiting to avoid some deadhangs (#25122)
Depending on how diagnostic notifications come in, we might not
be able to await the necessary times for the number of cells, creating
a deadcount. Changing to just waiting for the diagnostics to all be
populated fixes this, and is what we actually care about anyways.

MODULAR_ORIG_COMMIT_REV_ID: a7fd54a5f4d80d04b288b8a340478d08247b29cc