[vscode] Fix the debug hook in the document's menu (#25113)
A while ago Mr River added a debug action in the current document's
menu. This patch fixes it giving the debug menu action a way to let the
debug configuration resolver know of the current SDK. Now there's a
launch and attach debug setting called modularHomePath for this. What is
nice of this setting is that the user can specify it in launch.json
files if they need to override settings or as a workaround.

The only issue left to resolve is
[Internal link] which causes some JIT
invocations to make LLDB fall in an infinite loop, but I'm going to
solve it soon.

I also did some minor cleanups here and there.

MODULAR_ORIG_COMMIT_REV_ID: 32e4915b8bb0210c7ec3decbe24e7c2903fdcb19