[Mojo] Record Captures After Signature Resolution (#25074)
If we are resolving the signature of a nested function, the declScope of
the emitter is the parent, not the nested function. As a result the
capture code is ignored. In fact, the DeclRefNode::emitIR is not the
appropriate place to record this capture because we do not have enough
information yet because we are still parsing (is it escaping?). Instead,
finish emitting the signature and then record captures afterwards.

MODULAR_ORIG_COMMIT_REV_ID: 8bfa3cf85b0f2441b24576437df35f3d5c4a83c8