[KGEN] Remove `kgen.link` path use case (#24944)
The `kgen.link` op currently accepts one of two arguments: either a
string attribute representing the path to a static library, or bytes
representing the object or library's contents.

The latter is used for `.mojopkg` Mojo packages, and the former isn't
used at all. It was first introduced in #14519, as part of supporting
issue #14325, which explains that it "could be useful to write a
function in C or assembly and be able to call it from Mojo." Undeniably
this could be useful, but a Mojo syntax that made it possible was never
added, and so it's not currently used.

Remove this use case for `kgen.link` for now; we can restore it should
it become useful again. In the meantime, removing this unused use case
makes it simpler to modify `kgen.link` to pull in transitive
dependencies, to address #22108.

MODULAR_ORIG_COMMIT_REV_ID: cd042179af849e56ee370641c83e9dacf5b63d3e