[Mojo] Fix Mangle Error By Preserving the Parameter Name (#25098)
Previously the integration test in this example did not compile because
the signature types were not mapping to the same closure wrapper type
due to differences in the parameter names. This patch fixes that by
persisting source names so the source name is used as the parameter name
instead of the mangled name.

MODULAR_ORIG_COMMIT_REV_ID: fdc32dec43ee3b4d69205f0e47a026358aa30ca5