[CUDA] Forgive mark_ready on chains in GPU-via-CPU kernels (#25209)
We have a lot of
```
@parameter
if not single_thread_blocking_override:
  out_chain.mark_ready()
```
hackery in mojo kernel preludes. Currently GPU kernels are expected to
be synchronous, with just the GPU portion launched as asynchronous using
the CUDA machinery, so the mark_ready would assert fail when emplacing
the underlying chain. Silently ignore these unnecessary mark_ready calls
to avoid churn.

Both the single_thread_blocking_override, and handling of the
async<->output_chain is not long for this world, so this quick fix seems
fine.

This allows
All/test/top-models-with-cuda/msft-baby-llama-with-cuda.smoke to run, so
I'll switch the test on, but the all-zeros output looks highly
suspicious.

MODULAR_ORIG_COMMIT_REV_ID: 2b41087eb31627b4688c155fe63bf5de278f8f33