[KGEN] Delay DebugInfo Value concretization until location elaboration (#25218)
Fixes #24872.

An edge case in the Elaborator occurs when we have a ParamIf, with a
ParamFork inside, and a DebugInfo::ValueOp outside, like this:
```MLIR
kgen.generator @A(%arg0) {
  debuginfo.value #local_variable = %arg0: index loc(#loc)
  kgen.param.if <1> {
    kgen.param.fork
    ...
}
```
When we process the fork op and clone the parent ImplNode, we replace
SubProgramScope for all the ops in the func. This is supposed to match
everything that previously belonged to this generator, but for the
debuginfo.value outside, the type of the subprogram of its
`#local_variable` has already been concretized. This causes a mismatch
between the scope in `#local_variable` and what the replacer is looking
for.

The real problem is not with the fork behavior, but with the eager
concretization of the debuginfo in debuginfo.value. At the time of
processing that op, we do not yet know the final name of the function it
will belong to (e.g. due to forks). We can, however, do the same thing
as how we handle location concretization, and do it at the end of
processing each ImplNode, which is what this PR does. This is OK because
any graph with DebugInfo already requires elaborate-locations to be
true, so we're not losing any expected behavior.

MODULAR_ORIG_COMMIT_REV_ID: dbab290fdd1a4c9d222401cdb939dd7ebab46f64