[Kernels][GPU] Enable slice and transpose on GPU (#25333)
This is super ugly and it might be better to wait for
[Internal link] But it at least gets
the ops running on GPU.

We also need to figure out a better strategy for transpose, it is
currently implmented as a view op but that will not work well on GPUs
because you will not get memory coalescing. This may be able to wait
until after modcon because all of the transposes in llama2 are handled
in the flash attention kernel.

Co-authored-by: Ian Tramble <ian@modular.com>
MOGG_ORIG_REV_ID: 5188d43de6e9deaad8cca8ee1a67fdfee00a26ea