[driver-tblgen] Add `CommandOption` multiclass (#25337)
The `CommandOption` multiclass allows for the definition of both
`--debug-level <level>` and `--debug-level=<level>` using a single
`defm` statement in TableGen -- in other words, the definition of an
option as well as an alias for that option that takes an `=` suffix.

The `=` aliases introduce too much clutter to help text, so they're
hidden. To support this, this patch also introduces support for
"hidden" options in driver-tblgen (but not hidden groups, or other
constructs). It's not the goal of this patch to introduce hidden options
to the `mojo` and `modular` command line tools, but this does make that
technically possible.

This adds the tblgen implementation necessary for #18424.

MODULAR_ORIG_COMMIT_REV_ID: 9aae4b53b08807cb2a2a5f2863159d8909c8b268