[mojo-lang] Simplify wrapper signature gen with parameters (NFC) (#25405)
Leverage `LITSignatureType::prependParams` to generate the new
signature. This slightly alters the assumptions when working with
parameter references. The assumption is that if a ParamDeclRefAttr can
be found within a name scope, a corresponding ParamDeclAttr must exist
for it somewhere (since we assume the parser emits correct code), and
thus it can be manipulated without directly checking for a declaration.

MODULAR_ORIG_COMMIT_REV_ID: b06fa89dcb314240bce61e61ad95d0ba6087ab5c