[mojo-lang] Restructure how parameter captures are handled (#25407)
This started with me trying to fix what seemed like a small bug in
closure emission that was emitting a `lit.call` op whose result type
didn't match the expected callee result type (`lit.call` was not
verifying the result type before).

The bug turned out to be related to the way parameter captures are
detected and propagated to the closure impl and wrapper structs, and
their methods.

This PR restructures how parameter captures are found and propagated,
relying on the assumption that if a `ParamDeclRefAttr` is found in the
IR, it must be valid and the declaration is available in the current
name scope. That means it doesn't matter where the parameter comes from:
at the capture site, we can hoist the nested function body out without
changing the parameter references, placing the new decls on the closure
impl struct, and then binding the parameters to closure impl and wrapper
structs as appropriate.

This PR also tweaks some handling of index/nested references to be more
robust.

MODULAR_ORIG_COMMIT_REV_ID: 7c3b25a53c784e919132d93469bba9b7b90d0709